requirements.txt
langchain==0.3.0
langchain_community==0.3.0
langchain_openai==0.2.0
python-dotenv==1.0.1
streamlit==1.38.0
faiss-cpu
